Move http responses to own package
